add function getDataWithValueSum
